requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.1.0
pytesseract==0.3.10
numpy==1.26.1
opencv-python==4.8.1.78
spacy==3.7.2
en-core-sci-sm @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_core_sci_sm-0.5.3.tar.gz
pydantic==2.4.2
pydantic-settings==2.0.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
openai==1.3.0
python-dotenv==1.0.0